PSY-800: move entity domain forms from components/forms into features
Relocate entity-domain form components out of frontend/components/forms/
into their owning feature modules (git mv, history preserved):

- features/artists/components/: ArtistEditForm, ArtistInput (+tests)
- features/shows/components/: ShowForm, show-form-utils,
  OrphanedArtistsDialog (+tests)
- features/venues/components/: VenueEditForm, VenueInput,
  venue-edit-utils (+tests)

OrphanedArtistsDialog goes to features/shows: its sole consumer is
ShowForm and its data type (OrphanedArtist) lives in the shows feature.

Within-feature imports converted to relative paths per feature-module
convention (no self-barrel imports); cross-feature consumers now import
via feature barrels. Test mocks re-pointed at the new module paths.
components/forms/ keeps only FormField + AIFormFiller (classification
deferred to PSY-799).
