+++ frontend/app/shows/submit/page.tsx
@@ -20,7 +20,8 @@ import {
   CardTitle,
 } from '@/components/ui/card'
 import { Button } from '@/components/ui/button'
-import { ShowForm, AIFormFiller } from '@/components/forms'
+import { AIFormFiller } from '@/components/forms'
+import { ShowForm } from '@/features/shows'
 import type { ExtractedShowData } from '@/lib/types/extraction'
 
 function EmailVerificationRequired() {
