+++ frontend/features/artists/index.ts
@@ -65,6 +65,8 @@ export { useReducedMotion } from './hooks'
 // Re-adding it here re-hoists ArtistDetail.tsx (~40 KB) into the shared chunk.
 export {
   ArtistCard,
+  ArtistEditForm,
+  ArtistInput,
   ArtistSearch,
   ArtistList,
   ArtistListSkeleton,
