+++ frontend/features/shows/components/OrphanedArtistsDialog.tsx
@@ -3,7 +3,7 @@
 import { useState } from 'react'
 import { Loader2, Trash2, UserX } from 'lucide-react'
 import { apiRequest, API_ENDPOINTS } from '@/lib/api'
-import type { OrphanedArtist } from '@/features/shows'
+import type { OrphanedArtist } from '../types'
 import {
   Dialog,
   DialogContent,
