+++ frontend/features/shows/components/index.ts
@@ -4,6 +4,7 @@ export { ShowDetail } from './ShowDetail'
 export { ShowHeader } from './ShowHeader'
 export { ShowActions } from './ShowActions'
 export { ShowCard } from './ShowCard'
+export { ShowForm } from './ShowForm'
 export type { ShowCardDensity, ShowCardProps } from './ShowCard'
 export { ShowList } from './ShowList'
 export { ShowListSkeleton } from './ShowListSkeleton'
