+++ frontend/features/shows/index.ts
@@ -118,6 +118,7 @@ export {
   AttendanceButton,
   ShowDetail,
   ShowCard,
+  ShowForm,
   ShowList,
   ShowListSkeleton,
   HomeShowList,
