+++ frontend/features/venues/components/VenueCard.tsx
@@ -16,9 +16,8 @@ import { useAuthContext } from '@/lib/context/AuthContext'
 import { useQueryClient } from '@tanstack/react-query'
 import { createInvalidateQueries } from '@/lib/queryClient'
 import type { VenueWithShowCount } from '../types'
-import { ShowForm } from '@/components/forms/ShowForm'
-import { VenueEditForm } from '@/components/forms/VenueEditForm'
-import { CompactShowRow, SHOW_LIST_FEATURE_POLICY } from '@/features/shows'
+import { CompactShowRow, SHOW_LIST_FEATURE_POLICY, ShowForm } from '@/features/shows'
+import { VenueEditForm } from './VenueEditForm'
 import { DeleteVenueDialog } from './DeleteVenueDialog'
 import { Button } from '@/components/ui/button'
