+++ frontend/features/venues/components/VenueEditForm.tsx
@@ -3,9 +3,16 @@
 import { useState } from 'react'
 import { useForm } from '@tanstack/react-form'
 import { z } from 'zod'
-import { Loader2, Edit2, AlertCircle, CheckCircle2 } from 'lucide-react'
-import { useArtistUpdate } from '@/lib/hooks/admin/useAdminArtists'
-import type { Artist, ArtistEditRequest } from '@/features/artists'
+import {
+  Edit2,
+  AlertCircle,
+  CheckCircle2,
+  Loader2,
+} from 'lucide-react'
+import { useVenueUpdate } from '../hooks/useVenueEdit'
+import { useAuthContext } from '@/lib/context/AuthContext'
+import type { VenueWithShowCount, Venue } from '../types'
+import { detectVenueChanges, type VenueEditFormValues } from './venue-edit-utils'
 import {
   Dialog,
   DialogContent,
@@ -18,12 +25,15 @@ import { Button } from '@/components/ui/button'
 import { Input } from '@/components/ui/input'
 import { Label } from '@/components/ui/label'
 import { Alert, AlertDescription } from '@/components/ui/alert'
-import { FieldInfo } from './FormField'
+import { FieldInfo } from '@/components/forms'
 
-const artistEditSchema = z.object({
-  name: z.string().min(1, 'Artist name is required'),
-  city: z.string(),
-  state: z.string(),
+// Form validation schema
+const venueEditSchema = z.object({
+  name: z.string().min(1, 'Venue name is required'),
+  address: z.string(),
+  city: z.string().min(1, 'City is required'),
+  state: z.string().min(2, 'State is required'),
+  zipcode: z.string(),
   instagram: z.string(),
   facebook: z.string(),
   twitter: z.string(),
@@ -34,38 +44,30 @@ const artistEditSchema = z.object({
   website: z.string(),
 })
 
-interface FormValues {
-  name: string
-  city: string
-  state: string
-  instagram: string
-  facebook: string
-  twitter: string
-  youtube: string
-  spotify: string
-  soundcloud: string
-  bandcamp: string
-  website: string
-}
+type FormValues = VenueEditFormValues
 
-interface ArtistEditFormProps {
-  artist: Artist
+interface VenueEditFormProps {
+  venue: VenueWithShowCount | Venue
   open: boolean
   onOpenChange: (open: boolean) => void
   onSuccess?: () => void
 }
 
-// Callers MUST pass `key={artist.id}` so React unmounts + remounts with
-// fresh state when the artist switches. The form deliberately does NOT
+// Admin-only direct-edit form. Non-admin edits go through the unified
+// suggest-edit flow (EntityEditDrawer / useSuggestEdit).
+//
+// Callers MUST pass `key={venue.id}` so React unmounts + remounts with
+// fresh state when the venue switches. The form deliberately does NOT
 // useEffect to reset prop-derived state — see React's "You Might Not
 // Need an Effect" guide for the rationale.
-export function ArtistEditForm({
-  artist,
+export function VenueEditForm({
+  venue,
   open,
   onOpenChange,
   onSuccess,
-}: ArtistEditFormProps) {
-  const updateMutation = useArtistUpdate()
+}: VenueEditFormProps) {
+  const { user } = useAuthContext()
+  const updateMutation = useVenueUpdate()
   const [showSuccess, setShowSuccess] = useState(false)
   const [error, setError] = useState<string | null>(null)
 
@@ -74,18 +76,23 @@ export function ArtistEditForm({
     setShowSuccess(false)
   }
 
+  const isAdmin = user?.is_admin ?? false
+
+  // Initialize form with venue data
   const initialValues: FormValues = {
-    name: artist.name,
-    city: artist.city || '',
-    state: artist.state || '',
-    instagram: artist.social?.instagram || '',
-    facebook: artist.social?.facebook || '',
-    twitter: artist.social?.twitter || '',
-    youtube: artist.social?.youtube || '',
-    spotify: artist.social?.spotify || '',
-    soundcloud: artist.social?.soundcloud || '',
-    bandcamp: artist.social?.bandcamp || '',
-    website: artist.social?.website || '',
+    name: venue.name,
+    address: venue.address || '',
+    city: venue.city,
+    state: venue.state,
+    zipcode: venue.zipcode || '',
+    instagram: venue.social?.instagram || '',
+    facebook: venue.social?.facebook || '',
+    twitter: venue.social?.twitter || '',
+    youtube: venue.social?.youtube || '',
+    spotify: venue.social?.spotify || '',
+    soundcloud: venue.social?.soundcloud || '',
+    bandcamp: venue.social?.bandcamp || '',
+    website: venue.social?.website || '',
   }
 
   const form = useForm({
@@ -93,38 +100,15 @@ export function ArtistEditForm({
     onSubmit: async ({ value }) => {
       setError(null)
 
-      // Build request with only changed fields
-      const changes: ArtistEditRequest = {}
+      const changes = detectVenueChanges(value, venue)
 
-      if (value.name !== artist.name) changes.name = value.name
-      if (value.city !== (artist.city || ''))
-        changes.city = value.city || undefined
-      if (value.state !== (artist.state || ''))
-        changes.state = value.state || undefined
-      if (value.instagram !== (artist.social?.instagram || ''))
-        changes.instagram = value.instagram || undefined
-      if (value.facebook !== (artist.social?.facebook || ''))
-        changes.facebook = value.facebook || undefined
-      if (value.twitter !== (artist.social?.twitter || ''))
-        changes.twitter = value.twitter || undefined
-      if (value.youtube !== (artist.social?.youtube || ''))
-        changes.youtube = value.youtube || undefined
-      if (value.spotify !== (artist.social?.spotify || ''))
-        changes.spotify = value.spotify || undefined
-      if (value.soundcloud !== (artist.social?.soundcloud || ''))
-        changes.soundcloud = value.soundcloud || undefined
-      if (value.bandcamp !== (artist.social?.bandcamp || ''))
-        changes.bandcamp = value.bandcamp || undefined
-      if (value.website !== (artist.social?.website || ''))
-        changes.website = value.website || undefined
-
-      if (Object.keys(changes).length === 0) {
+      if (!changes) {
         setError('No changes detected')
         return
       }
 
       updateMutation.mutate(
-        { artistId: artist.id, data: changes },
+        { venueId: venue.id, data: changes },
         {
           onSuccess: () => {
             setShowSuccess(true)
@@ -136,14 +120,14 @@ export function ArtistEditForm({
           },
           onError: err => {
             setError(
-              err instanceof Error ? err.message : 'Failed to update artist'
+              err instanceof Error ? err.message : 'Failed to update venue'
             )
           },
         }
       )
     },
     validators: {
-      onSubmit: artistEditSchema,
+      onSubmit: venueEditSchema,
     },
   })
 
@@ -154,16 +138,22 @@ export function ArtistEditForm({
     onOpenChange(nextOpen)
   }
 
+  // Non-admins should not see this form. Guard here as a safety net;
+  // VenueCard.canEdit should already hide the trigger for non-admins.
+  if (!isAdmin) {
+    return null
+  }
+
   return (
     <Dialog open={open} onOpenChange={handleDialogOpenChange}>
       <DialogContent className="sm:max-w-lg max-h-[90vh] overflow-y-auto">
         <DialogHeader>
           <DialogTitle className="flex items-center gap-2">
             <Edit2 className="h-5 w-5" />
-            Edit Artist
+            Edit Venue
           </DialogTitle>
           <DialogDescription>
-            Make changes to this artist. Changes will be applied immediately.
+            Make changes to this venue. Changes will be applied immediately.
           </DialogDescription>
         </DialogHeader>
 
@@ -171,7 +161,7 @@ export function ArtistEditForm({
           <Alert className="mb-4 border-success-foreground bg-success">
             <CheckCircle2 className="h-4 w-4 text-success-foreground" />
             <AlertDescription className="text-success-foreground">
-              Artist updated successfully!
+              Venue updated successfully!
             </AlertDescription>
           </Alert>
         )}
@@ -198,16 +188,31 @@ export function ArtistEditForm({
             <form.Field name="name">
               {field => (
                 <div className="space-y-2">
-                  <Label htmlFor="artist-name">Name *</Label>
+                  <Label htmlFor="name">Venue Name *</Label>
                   <Input
-                    id="artist-name"
+                    id="name"
                     value={field.state.value}
                     onChange={e => field.handleChange(e.target.value)}
                     onBlur={field.handleBlur}
-                    placeholder="e.g. The National"
+                    placeholder="e.g. The Empty Bottle"
                     aria-invalid={field.state.meta.errors.length > 0}
                   />
                   <FieldInfo field={field} />
+                </div>
+              )}
+            </form.Field>
+
+            <form.Field name="address">
+              {field => (
+                <div className="space-y-2">
+                  <Label htmlFor="address">Address</Label>
+                  <Input
+                    id="address"
+                    value={field.state.value}
+                    onChange={e => field.handleChange(e.target.value)}
+                    onBlur={field.handleBlur}
+                    placeholder="e.g. 1035 N Western Ave"
+                  />
                 </div>
               )}
             </form.Field>
@@ -216,14 +221,16 @@ export function ArtistEditForm({
               <form.Field name="city">
                 {field => (
                   <div className="space-y-2">
-                    <Label htmlFor="artist-city">City</Label>
+                    <Label htmlFor="city">City *</Label>
                     <Input
-                      id="artist-city"
+                      id="city"
                       value={field.state.value}
                       onChange={e => field.handleChange(e.target.value)}
                       onBlur={field.handleBlur}
-                      placeholder="e.g. Phoenix"
+                      placeholder="e.g. Chicago"
+                      aria-invalid={field.state.meta.errors.length > 0}
                     />
+                    <FieldInfo field={field} />
                   </div>
                 )}
               </form.Field>
@@ -231,18 +238,35 @@ export function ArtistEditForm({
               <form.Field name="state">
                 {field => (
                   <div className="space-y-2">
-                    <Label htmlFor="artist-state">State</Label>
+                    <Label htmlFor="state">State *</Label>
                     <Input
-                      id="artist-state"
+                      id="state"
                       value={field.state.value}
                       onChange={e => field.handleChange(e.target.value)}
                       onBlur={field.handleBlur}
-                      placeholder="e.g. AZ"
+                      placeholder="e.g. IL"
+                      aria-invalid={field.state.meta.errors.length > 0}
                     />
+                    <FieldInfo field={field} />
                   </div>
                 )}
               </form.Field>
             </div>
+
+            <form.Field name="zipcode">
+              {field => (
+                <div className="space-y-2">
+                  <Label htmlFor="zipcode">Zipcode</Label>
+                  <Input
+                    id="zipcode"
+                    value={field.state.value}
+                    onChange={e => field.handleChange(e.target.value)}
+                    onBlur={field.handleBlur}
+                    placeholder="e.g. 60622"
+                  />
+                </div>
+              )}
+            </form.Field>
           </div>
 
           {/* Social Links */}
@@ -253,9 +277,9 @@ export function ArtistEditForm({
               <form.Field name="website">
                 {field => (
                   <div className="space-y-2">
-                    <Label htmlFor="artist-website">Website</Label>
+                    <Label htmlFor="website">Website</Label>
                     <Input
-                      id="artist-website"
+                      id="website"
                       value={field.state.value}
                       onChange={e => field.handleChange(e.target.value)}
                       onBlur={field.handleBlur}
@@ -268,9 +292,9 @@ export function ArtistEditForm({
               <form.Field name="instagram">
                 {field => (
                   <div className="space-y-2">
-                    <Label htmlFor="artist-instagram">Instagram</Label>
+                    <Label htmlFor="instagram">Instagram</Label>
                     <Input
-                      id="artist-instagram"
+                      id="instagram"
                       value={field.state.value}
                       onChange={e => field.handleChange(e.target.value)}
                       onBlur={field.handleBlur}
@@ -283,9 +307,9 @@ export function ArtistEditForm({
               <form.Field name="facebook">
                 {field => (
                   <div className="space-y-2">
-                    <Label htmlFor="artist-facebook">Facebook</Label>
+                    <Label htmlFor="facebook">Facebook</Label>
                     <Input
-                      id="artist-facebook"
+                      id="facebook"
                       value={field.state.value}
                       onChange={e => field.handleChange(e.target.value)}
                       onBlur={field.handleBlur}
@@ -298,9 +322,9 @@ export function ArtistEditForm({
               <form.Field name="twitter">
                 {field => (
                   <div className="space-y-2">
-                    <Label htmlFor="artist-twitter">Twitter/X</Label>
+                    <Label htmlFor="twitter">Twitter/X</Label>
                     <Input
-                      id="artist-twitter"
+                      id="twitter"
                       value={field.state.value}
                       onChange={e => field.handleChange(e.target.value)}
                       onBlur={field.handleBlur}
@@ -310,27 +334,12 @@ export function ArtistEditForm({
                 )}
               </form.Field>
 
-              <form.Field name="youtube">
-                {field => (
-                  <div className="space-y-2">
-                    <Label htmlFor="artist-youtube">YouTube</Label>
-                    <Input
-                      id="artist-youtube"
-                      value={field.state.value}
-                      onChange={e => field.handleChange(e.target.value)}
-                      onBlur={field.handleBlur}
-                      placeholder="https://youtube.com/..."
-                    />
-                  </div>
-                )}
-              </form.Field>
-
               <form.Field name="spotify">
                 {field => (
                   <div className="space-y-2">
-                    <Label htmlFor="artist-spotify">Spotify</Label>
+                    <Label htmlFor="spotify">Spotify</Label>
                     <Input
-                      id="artist-spotify"
+                      id="spotify"
                       value={field.state.value}
                       onChange={e => field.handleChange(e.target.value)}
                       onBlur={field.handleBlur}
@@ -340,27 +349,12 @@ export function ArtistEditForm({
                 )}
               </form.Field>
 
-              <form.Field name="soundcloud">
-                {field => (
-                  <div className="space-y-2">
-                    <Label htmlFor="artist-soundcloud">SoundCloud</Label>
-                    <Input
-                      id="artist-soundcloud"
-                      value={field.state.value}
-                      onChange={e => field.handleChange(e.target.value)}
-                      onBlur={field.handleBlur}
-                      placeholder="https://soundcloud.com/..."
-                    />
-                  </div>
-                )}
-              </form.Field>
-
               <form.Field name="bandcamp">
                 {field => (
                   <div className="space-y-2">
-                    <Label htmlFor="artist-bandcamp">Bandcamp</Label>
+                    <Label htmlFor="bandcamp">Bandcamp</Label>
                     <Input
-                      id="artist-bandcamp"
+                      id="bandcamp"
                       value={field.state.value}
                       onChange={e => field.handleChange(e.target.value)}
                       onBlur={field.handleBlur}
