+++ frontend/features/venues/components/index.ts
@@ -4,6 +4,8 @@ export { VenueDetail } from './VenueDetail'
 export { VenueList } from './VenueList'
 export { VenueLocationCard } from './VenueLocationCard'
 export { VenueShowsList } from './VenueShowsList'
+export { VenueEditForm } from './VenueEditForm'
+export { VenueInput } from './VenueInput'
 export { VenueBillNetwork } from './VenueBillNetwork'
 export { DeleteVenueDialog } from './DeleteVenueDialog'
 export { VenueDeniedDialog } from './VenueDeniedDialog'
